Removed dead code and added error handling to all applicable classes. Fidelity-DCO-1.1-Signed-off-by: Robert Coords<robert.coords@fmr.com>(github:robertcoords)

+++ tests/com/alliancefoundry/tests/KafkaTests.java
@@ -52,7 +52,7 @@ public class KafkaTests {
 		KafkaSubscriber kafkaSubscriber = new KafkaSubscriber(topic);
 		kafkaSubscriber.consumeEvent();
 		
-		assertEquals("Should be 42", 42, event.getSequenceNumber());
+		assertEquals("Should be 42", 42, event.getSequenceNumber().intValue());
 	}
 	
 	@Test
@@ -87,7 +87,7 @@ public class KafkaTests {
 		KafkaSubscriber kafkaSubscriber = new KafkaSubscriber(topic);
 		kafkaSubscriber.consumeEvent();
 		
-		assertEquals("Should be 12", 12, event2.getSequenceNumber());
+		assertEquals("Should be 12", 12, event2.getSequenceNumber().intValue());
 		
 	}
 }
